SaveTheChange was working fine, but cant get it to work when I apply SaveTheChange through the metaclass
