+++ otree/export.py
@@ -19,7 +19,6 @@ from otree.models.session import Session
 from otree.models.subsession import BaseSubsession
 from otree.models.group import BaseGroup
 from otree.models.player import BasePlayer
-from otree.db.serializedfields import json_loads
 from otree.session import SessionConfig
 
 from otree.models_concrete import (
