+++ tests/settings.py
@@ -20,7 +20,8 @@ SECRET_KEY = 'zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz'
 
 AUTH_LEVEL = ''
 
-'''
+environ['DATABASE_URL'] = 'postgres://postgres@localhost/django_db'
+
 DATABASES = {
     'default': dj_database_url.config(
         default='sqlite:///' + os.path.join(BASE_DIR, 'db.sqlite3')
@@ -44,7 +45,7 @@ DATABASES = {
         }
     }
 }
-
+'''
 
 CREATE_DEFAULT_SUPERUSER = True
 ADMIN_USERNAME = 'admin'
